First official release
This library was adapted from the "Bounce" library for use with capacitive touch inputs. To 'quiet' the input, thresholds are used instead of time intervals. The user sets the threshold at which the input should be considered ON. A separate OFF threshold (of -10%) is automatically implemented to reduce noise near the threshold. The same time-based functions of the Bounce library are included: 'retrigger' works the same as 'rebounce' and duration is the same.